Refactor recipe modal code into a separate class
